+++ packages/project/src/serialize/to-app-state.ts
@@ -141,11 +141,12 @@ export const mapWorkflow = (
           });
           if (mappedCredential) {
             projectCredentialId = mappedCredential.uuid;
-          } else {
-            console.warn(`WARING! Failed to map credential ${projectCredentialId} - Lightning may throw an error.
-              
-Ensure the credential exists in project.yaml and try again (maybe ensure the credential is attached to the project in the app and run project fetch)`);
           }
+          //            else {
+          //             console.warn(`WARING! Failed to map credential ${projectCredentialId} - Lightning may throw an error.
+
+          // Ensure the credential exists in project.yaml and try again (maybe ensure the credential is attached to the project in the app and run project fetch)`);
+          //           }
           otherOpenFnProps.project_credential_id = projectCredentialId;
         }
       }